feat: move SQL relation extraction tests to dedicated files
